refactor: dragandDrop 함수를 kanban 컴포넌트까지 끌어올리기 [#87]
